diabetescheck/requirements-test.txt
-r requirements.txt

coverage==4.0.3
flake8==2.5.1
pytest==2.8.5
pytest-django==2.9.1
pytest-cov==2.2.0
model-mommy==1.2.6
tox